drivers: wifi: Store RPU firmware patches in the external flash
The internal flash utilization is almost full for Matter based
applications, so, free up memory by storing the firmware patches in the
external flash on supported platforms, for now only nRF5340 & nRF52840
are supported, as nRF7002 accesses external flash using SPI where XIP is
not supported.

Due to Nordic flash driver low-power optimizations,  we need to inform
the flash driver whenever XIP is being used, so, move the FW patch API
to a separate C file to avoid using other RODATA to minimize this XIP
management.

This is disabled by default till we get the necessary DFU changes in
Matter.
